workers/requirements.txt
httpx==0.27.0
pyyaml==6.0.1
kafka-python==2.0.2
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
prometheus-client==0.20.0
scikit-learn==1.4.2
numpy==1.26.4
requests==2.31.0
